Move SubHeader component and rename it to FormModalSubHeader
