+++ packages/core/content-type-builder/admin/src/components/FormModal/utils/index.js
@@ -1,4 +1,3 @@
 export { default as canEditContentType } from './canEditContentType';
 export { default as getAttributesToDisplay } from './getAttributesToDisplay';
-export { default as getModalTitleSubHeader } from './getModalTitleSubHeader';
 export { default as getFormInputNames } from './getFormInputNames';
